backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database and ORM
supabase==2.0.2
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1

# Vector search and embeddings
pgvector==0.2.4
openai==1.3.5
sentence-transformers==2.2.2
chromadb==0.4.18

# AI agents and LLM frameworks
crewai==0.1.55
langchain==0.0.340
langchain-openai==0.0.2
langchain-community==0.0.8
langgraph==0.0.20

# PDF processing
pypdf2==3.0.1
pdfplumber==0.10.0
pymupdf==1.23.8
reportlab==4.0.7

# File processing
python-magic==0.4.27
pillow==10.1.0
python-docx==1.1.0
openpyxl==3.1.2

# Email integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# HTTP and async
httpx==0.25.2
aiofiles==23.2.1
aiohttp==3.9.1

# Data processing
pandas==2.1.3
numpy==1.25.2
pydantic==2.5.0
pydantic-settings==2.0.3

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
redis==5.0.1
celery==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

